Set EditText input type from UxTweaksForwarder
