+++ app/src/main/java/fr/neamar/kiss/SettingsActivity.java
@@ -31,7 +31,7 @@ public class SettingsActivity extends PreferenceActivity implements
 
     // Those settings require the app to restart
 
-    final static private String requireRestartSettings = "enable-keyboard-workaround force-portrait primary-color transparent-search transparent-favorites history-hide enable-favorites-bar";
+    final static private String requireRestartSettings = "force-portrait primary-color transparent-search transparent-favorites history-hide enable-favorites-bar";
 
     final static private String requireInstantRestart = "theme notification-bar-color";
