+++ app/src/main/java/fr/neamar/kiss/forwarder/ForwarderManager.java
@@ -32,6 +32,7 @@ public class ForwarderManager extends Forwarder {
         favoriteForwarder.onCreate();
         widgetForwarder.onCreate();
         uiTweaksForwarder.onCreate();
+        uxTweaksForwarder.onCreate();
     }
 
     @Override
